se termino la nuva sección de la api rest

+++ app/Http/Controllers/JornadaController.php
@@ -4,12 +4,27 @@ namespace App\Http\Controllers;
 
 use App\Jornada;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class JornadaController extends Controller
 {
     public function index()
     {
-        $jornada = Jornada::all();
+        $empresa = Auth::user()->rfc;
+        $jornada = DB::table('jornadas')
+            ->where('jornadas.rfc_empresa', $empresa)
+            ->join('empresas', 'jornadas.rfc_empresa', 'empresas.rfc')
+            ->select(
+                'jornadas.id',
+                'jornadas.turno',
+                'jornadas.hora_inicial',
+                'jornadas.hora_final',
+                'jornadas.jornada',
+                'jornadas.rfc_empresa',
+                'empresas.nombre_empresa'
+            )
+            ->get();
         return response()->json([
             'total_jornadas' => count($jornada),
             'jornadas' => $jornada
@@ -20,6 +35,7 @@ class JornadaController extends Controller
         $this->validar($request);
         $datos = $request->all();
         $datos['jornada'] = 'fijo ' . $datos['turno'] . ' entre las ' . $datos['hora_inicial'] . ' y ' . $datos['hora_final'] . ' hrs ';
+        $datos['rfc_empresa'] = Auth::user()->rfc;
         $jornada = Jornada::create($datos);
         return response()->json([
             'Mensaje' => 'Se creo con exito la jornada.',
@@ -28,23 +44,49 @@ class JornadaController extends Controller
     }
     public function show($id)
     {
+        $empresa = Auth::user()->rfc;
         $jornada = Jornada::find($id);
         if (is_null($jornada)) {
             return response()->json([
                 'Error' => 'No se encontro la jornada con dicho id.'
             ], 404);
         }
+        if ($empresa != $jornada['rfc_empresa']) {
+            return response()->json([
+                'Error' => 'No se puede acceder a dicha información.'
+            ], 500);
+        }
+
+        $jornada = DB::table('jornadas')
+            ->where('jornadas.id', $id)
+            ->join('empresas', 'jornadas.rfc_empresa', 'empresas.rfc')
+            ->select(
+                'jornadas.id',
+                'jornadas.turno',
+                'jornadas.hora_inicial',
+                'jornadas.hora_final',
+                'jornadas.jornada',
+                'jornadas.rfc_empresa',
+                'empresas.nombre_nombre_empresa'
+            )
+            ->first();
         return response()->json([
-            'jornada' => $jornada,
+            'jornada' => $jornada
         ], 200);
     }
     public function update(Request $request, $id)
     {
+        $empresa = Auth::user()->rfc;
         $jornada = Jornada::find($id);
         if (is_null($jornada)) {
             return response()->json([
                 'Error' => 'No se encontro la jornada con dicho id.'
             ], 404);
+        }
+        if ($empresa != $jornada['rfc_empresa']) {
+            return response()->json([
+                'Error' => 'No se puede acceder a dicha información.'
+            ], 500);
         }
         $this->validar($request);
         $datos = $request->all();
@@ -57,11 +99,17 @@ class JornadaController extends Controller
     }
     public function destroy($id)
     {
+        $empresa = Auth::user()->rfc;
         $jornada = Jornada::find($id);
         if (is_null($jornada)) {
             return response()->json([
                 'Error' => 'No se encontro la jornada con dicho id.'
             ], 404);
+        }
+        if ($empresa != $jornada['rfc_empresa']) {
+            return response()->json([
+                'Error' => 'No se puede acceder a dicha información.'
+            ], 500);
         }
 
         $jornada->delete();
